@@ -16,16 +16,14 @@
 
 package com.example.android.observability.persistence;
 
-import android.arch.core.executor.testing.InstantTaskExecutorRule;
-import android.arch.persistence.room.Room;
-import android.support.test.InstrumentationRegistry;
+import androidx.arch.core.executor.testing.InstantTaskExecutorRule;
+import androidx.room.Room;
+import androidx.test.core.app.ApplicationProvider;
 
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
-
-import io.reactivex.functions.Predicate;
 
 /**
  * Integration tests for the {@link LocalUserDataSource} implementation with Room.
@@ -41,10 +39,10 @@ public class LocalUserDataSourceTest {
     private LocalUserDataSource mDataSource;
 
     @Before
-    public void initDb() throws Exception {
+    public void initDb() {
         // using an in-memory database because the information stored here disappears when the
         // process is killed
-        mDatabase = Room.inMemoryDatabaseBuilder(InstrumentationRegistry.getContext(),
+        mDatabase = Room.inMemoryDatabaseBuilder(ApplicationProvider.getApplicationContext(),
                 UsersDatabase.class)
                 // allowing main thread queries, just for testing
                 .allowMainThreadQueries()
@@ -53,56 +51,50 @@ public class LocalUserDataSourceTest {
     }
 
     @After
-    public void closeDb() throws Exception {
+    public void closeDb() {
         mDatabase.close();
     }
 
     @Test
     public void insertAndGetUser() {
         // When inserting a new user in the data source
-        mDataSource.insertOrUpdateUser(USER);
+        mDataSource.insertOrUpdateUser(USER).blockingAwait();
 
         // When subscribing to the emissions of the user
         mDataSource.getUser()
                 .test()
                 // assertValue asserts that there was only one emission of the user
-                .assertValue(new Predicate<User>() {
-                    @Override
-                    public boolean test(User user) throws Exception {
-                        // The emitted user is the expected one
-                        return user != null && user.getId().equals(USER.getId()) &&
-                                user.getUserName().equals(USER.getUserName());
-                    }
+                .assertValue(user -> {
+                    // The emitted user is the expected one
+                    return user != null && user.getId().equals(USER.getId()) &&
+                            user.getUserName().equals(USER.getUserName());
                 });
     }
 
     @Test
     public void updateAndGetUser() {
         // Given that we have a user in the data source
-        mDataSource.insertOrUpdateUser(USER);
+        mDataSource.insertOrUpdateUser(USER).blockingAwait();
 
         // When we are updating the name of the user
         User updatedUser = new User(USER.getId(), "new username");
-        mDataSource.insertOrUpdateUser(updatedUser);
+        mDataSource.insertOrUpdateUser(updatedUser).blockingAwait();
 
         // When subscribing to the emissions of the user
         mDatabase.userDao().getUser()
                 .test()
                 // assertValue asserts that there was only one emission of the user
-                .assertValue(new Predicate<User>() {
-                    @Override
-                    public boolean test(User user) throws Exception {
-                        // The emitted user is the expected one
-                        return user != null && user.getId().equals(USER.getId()) &&
-                                user.getUserName().equals("new username");
-                    }
+                .assertValue(user -> {
+                    // The emitted user is the expected one
+                    return user != null && user.getId().equals(USER.getId()) &&
+                            user.getUserName().equals("new username");
                 });
     }
 
     @Test
     public void deleteAndGetUser() {
         // Given that we have a user in the data source
-        mDataSource.insertOrUpdateUser(USER);
+        mDataSource.insertOrUpdateUser(USER).blockingAwait();
 
         //When we are deleting all users
         mDataSource.deleteAllUsers();
