@@ -16,29 +16,24 @@
 
 package com.example.android.observability.ui;
 
-import android.arch.lifecycle.LifecycleActivity;
-import android.arch.lifecycle.ViewModelProviders;
+import androidx.lifecycle.ViewModelProviders;
 import android.os.Bundle;
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
-import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.TextView;
-
 import com.example.android.observability.Injection;
 import com.example.android.persistence.R;
-
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.disposables.CompositeDisposable;
-import io.reactivex.functions.Action;
-import io.reactivex.functions.Consumer;
 import io.reactivex.schedulers.Schedulers;
 
 
 /**
  * Main screen of the app. Displays a user name and gives the option to update the user name.
  */
-public class UserActivity extends LifecycleActivity {
+public class UserActivity extends AppCompatActivity {
 
     private static final String TAG = UserActivity.class.getSimpleName();
 
@@ -59,18 +54,13 @@ public class UserActivity extends LifecycleActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_user);
 
-        mUserName = (TextView) findViewById(R.id.user_name);
-        mUserNameInput = (EditText) findViewById(R.id.user_name_input);
-        mUpdateButton = (Button) findViewById(R.id.update_user);
+        mUserName = findViewById(R.id.user_name);
+        mUserNameInput = findViewById(R.id.user_name_input);
+        mUpdateButton = findViewById(R.id.update_user);
 
         mViewModelFactory = Injection.provideViewModelFactory(this);
         mViewModel = ViewModelProviders.of(this, mViewModelFactory).get(UserViewModel.class);
-        mUpdateButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                updateUserName();
-            }
-        });
+        mUpdateButton.setOnClickListener(v -> updateUserName());
     }
 
     @Override
@@ -82,17 +72,8 @@ public class UserActivity extends LifecycleActivity {
         mDisposable.add(mViewModel.getUserName()
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
-                .subscribe(new Consumer<String>() {
-                    @Override
-                    public void accept(String userName) throws Exception {
-                        mUserName.setText(userName);
-                    }
-                }, new Consumer<Throwable>() {
-                    @Override
-                    public void accept(Throwable throwable) throws Exception {
-                        Log.e(TAG, "Unable to update username", throwable);
-                    }
-                }));
+                .subscribe(userName -> mUserName.setText(userName),
+                        throwable -> Log.e(TAG, "Unable to update username", throwable)));
     }
 
     @Override
@@ -112,16 +93,7 @@ public class UserActivity extends LifecycleActivity {
         mDisposable.add(mViewModel.updateUserName(userName)
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
-                .subscribe(new Action() {
-                    @Override
-                    public void run() throws Exception {
-                        mUpdateButton.setEnabled(true);
-                    }
-                }, new Consumer<Throwable>() {
-                    @Override
-                    public void accept(Throwable throwable) throws Exception {
-                        Log.e(TAG, "Unable to update username", throwable);
-                    }
-                }));
+                .subscribe(() -> mUpdateButton.setEnabled(true),
+                        throwable -> Log.e(TAG, "Unable to update username", throwable)));
     }
 }
