@@ -18,12 +18,10 @@ package com.example.android.contentprovidersample;
 
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
-
-import android.arch.persistence.room.Room;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.filters.SmallTest;
-import android.support.test.runner.AndroidJUnit4;
-
+import androidx.room.Room;
+import androidx.test.core.app.ApplicationProvider;
+import androidx.test.filters.SmallTest;
+import androidx.test.runner.AndroidJUnit4;
 import com.example.android.contentprovidersample.data.Cheese;
 import com.example.android.contentprovidersample.data.SampleDatabase;
 
@@ -43,7 +41,7 @@ public class CheeseTest {
 
     @Before
     public void createDatabase() {
-        mDatabase = Room.inMemoryDatabaseBuilder(InstrumentationRegistry.getTargetContext(),
+        mDatabase = Room.inMemoryDatabaseBuilder(ApplicationProvider.getApplicationContext()(),
                 SampleDatabase.class).build();
     }
 
