@@ -18,12 +18,12 @@ package com.example.android.contentprovidersample;
 
 import android.database.Cursor;
 import android.os.Bundle;
-import android.support.v4.app.LoaderManager;
-import android.support.v4.content.CursorLoader;
-import android.support.v4.content.Loader;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.loader.app.LoaderManager;
+import androidx.loader.content.CursorLoader;
+import androidx.loader.content.Loader;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.ViewGroup;
 import android.widget.TextView;
@@ -49,7 +49,7 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.main_activity);
 
-        final RecyclerView list = (RecyclerView) findViewById(R.id.list);
+        final RecyclerView list = findViewById(R.id.list);
         list.setLayoutManager(new LinearLayoutManager(list.getContext()));
         mCheeseAdapter = new CheeseAdapter();
         list.setAdapter(mCheeseAdapter);
@@ -127,7 +127,7 @@ public class MainActivity extends AppCompatActivity {
             ViewHolder(ViewGroup parent) {
                 super(LayoutInflater.from(parent.getContext()).inflate(
                         android.R.layout.simple_list_item_1, parent, false));
-                mText = (TextView) itemView.findViewById(android.R.id.text1);
+                mText = itemView.findViewById(android.R.id.text1);
             }
 
         }
