@@ -21,15 +21,14 @@ import static com.example.android.persistence.migrations.UsersDatabase.MIGRATION
 
 import static org.junit.Assert.assertEquals;
 
-import android.arch.persistence.db.SupportSQLiteDatabase;
-import android.arch.persistence.db.framework.FrameworkSQLiteOpenHelperFactory;
-import android.arch.persistence.room.Room;
-import android.arch.persistence.room.testing.MigrationTestHelper;
+import androidx.sqlite.db.SupportSQLiteDatabase;
+import androidx.sqlite.db.framework.FrameworkSQLiteOpenHelperFactory;
+import androidx.room.Room;
+import androidx.room.testing.MigrationTestHelper;
 import android.content.ContentValues;
 import android.database.sqlite.SQLiteDatabase;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
-
+import androidx.test.core.app.ApplicationProvider;
+import androidx.test.runner.AndroidJUnit4;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
@@ -64,7 +63,7 @@ public class MigrationTest {
     public void setUp() throws Exception {
         // To test migrations from version 1 of the database, we need to create the database
         // with version 1 using SQLite API
-        mSqliteTestDbHelper = new SqliteTestDbOpenHelper(InstrumentationRegistry.getTargetContext(),
+        mSqliteTestDbHelper = new SqliteTestDbOpenHelper(ApplicationProvider.getApplicationContext()(),
                 TEST_DB_NAME);
         // We're creating the table for every test, to ensure that the table is in the correct state
         SqliteDatabaseTestHelper.createTable(mSqliteTestDbHelper);
@@ -137,7 +136,7 @@ public class MigrationTest {
     }
 
     private UsersDatabase getMigratedRoomDatabase() {
-        UsersDatabase database = Room.databaseBuilder(InstrumentationRegistry.getTargetContext(),
+        UsersDatabase database = Room.databaseBuilder(ApplicationProvider.getApplicationContext()(),
                 UsersDatabase.class, TEST_DB_NAME)
                 .addMigrations(MIGRATION_1_2, MIGRATION_2_3)
                 .build();
