@@ -18,8 +18,7 @@ package com.example.android.persistence.migrations;
 
 import android.os.Handler;
 import android.os.Looper;
-import android.support.annotation.NonNull;
-
+import androidx.annotation.NonNull;
 import java.util.concurrent.Executor;
 import java.util.concurrent.Executors;
 
