@@ -17,7 +17,7 @@
 package com.example.android.persistence.migrations;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
@@ -41,16 +41,13 @@ public class UserActivity extends AppCompatActivity implements UserView {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_user);
 
-        mUserName = (TextView) findViewById(R.id.user_name);
-        mUserNameInput = (EditText) findViewById(R.id.user_name_input);
-        mUpdateButton = (Button) findViewById(R.id.update_user);
+        mUserName = findViewById(R.id.user_name);
+        mUserNameInput = findViewById(R.id.user_name_input);
+        mUpdateButton = findViewById(R.id.update_user);
 
-        mUpdateButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                String userName = mUserNameInput.getText().toString();
-                mPresenter.updateUserName(userName);
-            }
+        mUpdateButton.setOnClickListener(v -> {
+            String userName = mUserNameInput.getText().toString();
+            mPresenter.updateUserName(userName);
         });
 
         // Creating the repository here for simplicity.
