@@ -16,8 +16,7 @@
 
 package com.example.android.persistence.migrations;
 
-import android.support.annotation.NonNull;
-
+import androidx.annotation.NonNull;
 import java.util.concurrent.Executor;
 
 /**
@@ -27,12 +26,7 @@ import java.util.concurrent.Executor;
  * android.arch.core:core-testing to your build.gradle file.
  */
 public class SingleExecutors extends AppExecutors {
-    private static Executor instant = new Executor() {
-        @Override
-        public void execute(@NonNull Runnable command) {
-            command.run();
-        }
-    };
+    private static Executor instant = command -> command.run();
 
     public SingleExecutors() {
         super(instant, instant, instant);
