@@ -16,11 +16,11 @@
 
 package com.example.android.observability.persistence;
 
-import android.arch.persistence.room.ColumnInfo;
-import android.arch.persistence.room.Entity;
-import android.arch.persistence.room.Ignore;
-import android.arch.persistence.room.PrimaryKey;
-
+import androidx.room.ColumnInfo;
+import androidx.room.Entity;
+import androidx.room.Ignore;
+import androidx.room.PrimaryKey;
+import androidx.annotation.NonNull;
 import java.util.UUID;
 
 /**
@@ -29,6 +29,7 @@ import java.util.UUID;
 @Entity(tableName = "users")
 public class User {
 
+    @NonNull
     @PrimaryKey
     @ColumnInfo(name = "userid")
     private String mId;
