@@ -16,16 +16,12 @@
 
 package com.example.android.observability.ui;
 
-import android.arch.lifecycle.ViewModel;
-
 import com.example.android.observability.UserDataSource;
 import com.example.android.observability.persistence.User;
 
+import androidx.lifecycle.ViewModel;
 import io.reactivex.Completable;
 import io.reactivex.Flowable;
-import io.reactivex.functions.Action;
-import io.reactivex.functions.Function;
-import io.reactivex.internal.operators.completable.CompletableFromAction;
 
 /**
  * View Model for the {@link UserActivity}
@@ -48,12 +44,9 @@ public class UserViewModel extends ViewModel {
     public Flowable<String> getUserName() {
         return mDataSource.getUser()
                 // for every emission of the user, get the user name
-                .map(new Function<User, String>() {
-                    @Override
-                    public String apply(User user) throws Exception {
-                        mUser = user;
-                        return user.getUserName();
-                    }
+                .map(user -> {
+                    mUser = user;
+                    return user.getUserName();
                 });
 
     }
@@ -65,18 +58,12 @@ public class UserViewModel extends ViewModel {
      * @return a {@link Completable} that completes when the user name is updated
      */
     public Completable updateUserName(final String userName) {
-        return new CompletableFromAction(new Action() {
-            @Override
-            public void run() throws Exception {
-                // if there's no use, create a new user.
-                // if we already have a user, then, since the user object is immutable,
-                // create a new user, with the id of the previous user and the updated user name.
-                mUser = mUser == null
-                        ? new User(userName)
-                        : new User(mUser.getId(), userName);
-
-                mDataSource.insertOrUpdateUser(mUser);
-            }
-        });
+        // if there's no user, create a new user.
+        // if we already have a user, then, since the user object is immutable,
+        // create a new user, with the id of the previous user and the updated user name.
+        mUser = mUser == null
+                ? new User(userName)
+                : new User(mUser.getId(), userName);
+        return mDataSource.insertOrUpdateUser(mUser);
     }
 }
