@@ -16,10 +16,11 @@
 
 package com.example.android.observability.persistence
 
-import android.arch.persistence.room.Dao
-import android.arch.persistence.room.Insert
-import android.arch.persistence.room.OnConflictStrategy
-import android.arch.persistence.room.Query
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
+import io.reactivex.Completable
 
 import io.reactivex.Flowable
 
@@ -43,7 +44,7 @@ interface UserDao {
      * @param user the user to be inserted.
      */
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    fun insertUser(user: User)
+    fun insertUser(user: User): Completable
 
     /**
      * Delete all users.
