@@ -16,13 +16,11 @@
 
 package com.example.android.observability.ui
 
-import android.arch.lifecycle.ViewModel
+import androidx.lifecycle.ViewModel
 import com.example.android.observability.persistence.User
 import com.example.android.observability.persistence.UserDao
 import io.reactivex.Completable
 import io.reactivex.Flowable
-import io.reactivex.functions.Action
-import io.reactivex.internal.operators.completable.CompletableFromAction
 
 /**
  * View Model for the [UserActivity]
@@ -47,10 +45,8 @@ class UserViewModel(private val dataSource: UserDao) : ViewModel() {
      * @return a [Completable] that completes when the user name is updated
      */
     fun updateUserName(userName: String): Completable {
-        return CompletableFromAction(Action {
-            val user = User(USER_ID, userName)
-            dataSource.insertUser(user)
-        })
+        val user = User(USER_ID, userName)
+        return dataSource.insertUser(user)
     }
 
     companion object {
