@@ -16,13 +16,12 @@
 
 package com.example.android.observability
 
-import android.arch.core.executor.testing.InstantTaskExecutorRule
+import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import com.example.android.observability.persistence.User
 import com.example.android.observability.persistence.UserDao
 import com.example.android.observability.ui.UserViewModel
+import io.reactivex.Completable
 import io.reactivex.Flowable
-import org.hamcrest.MatcherAssert.assertThat
-import org.hamcrest.Matchers
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
@@ -30,29 +29,34 @@ import org.mockito.ArgumentCaptor
 import org.mockito.Captor
 import org.mockito.Mock
 import org.mockito.Mockito.`when`
-import org.mockito.Mockito.verify
 import org.mockito.MockitoAnnotations
+
 
 /**
  * Unit test for [UserViewModel]
  */
 class UserViewModelTest {
 
-    @get:Rule var instantTaskExecutorRule = InstantTaskExecutorRule()
+    @get:Rule
+    var instantTaskExecutorRule = InstantTaskExecutorRule()
 
-    @Mock private lateinit var dataSource: UserDao
+    @Mock
+    private lateinit var dataSource: UserDao
 
-    @Captor private lateinit var userArgumentCaptor: ArgumentCaptor<User>
+    @Captor
+    private lateinit var userArgumentCaptor: ArgumentCaptor<User>
 
     private lateinit var viewModel: UserViewModel
 
-    @Before fun setUp() {
+    @Before
+    fun setUp() {
         MockitoAnnotations.initMocks(this)
 
         viewModel = UserViewModel(dataSource)
     }
 
-    @Test fun getUserName_whenNoUserSaved() {
+    @Test
+    fun getUserName_whenNoUserSaved() {
         // Given that the UserDataSource returns an empty list of users
         `when`(dataSource.getUserById(UserViewModel.USER_ID)).thenReturn(Flowable.empty<User>())
 
@@ -63,7 +67,8 @@ class UserViewModelTest {
                 .assertNoValues()
     }
 
-    @Test fun getUserName_whenUserSaved() {
+    @Test
+    fun getUserName_whenUserSaved() {
         // Given that the UserDataSource returns a user
         val user = User(userName = "user name")
         `when`(dataSource.getUserById(UserViewModel.USER_ID)).thenReturn(Flowable.just(user))
@@ -75,17 +80,20 @@ class UserViewModelTest {
                 .assertValue("user name")
     }
 
-    @Test fun updateUserName_updatesNameInDataSource() {
+    @Test
+    fun updateUserName_updatesNameInDataSource() {
+        // Given that a user is already inserted
+        dataSource.insertUser(User(UserViewModel.USER_ID, "name"))
+
+        // And a specific user is expected when inserting
+        val userName = "new user name"
+        val expectedUser = User(UserViewModel.USER_ID, userName)
+        `when`(dataSource.insertUser(expectedUser)).thenReturn(Completable.complete())
+
         // When updating the user name
-        viewModel.updateUserName("new user name")
+        viewModel.updateUserName(userName)
                 .test()
                 .assertComplete()
-
-        // The user name is updated in the data source
-        // using ?: User("someUser") because otherwise, we get
-        // "IllegalStateException: userArgumentCaptor.capture() must not be null"
-        verify<UserDao>(dataSource).insertUser(capture(userArgumentCaptor))
-        assertThat(userArgumentCaptor.value.userName, Matchers.`is`("new user name"))
     }
 
 }
