@@ -16,13 +16,11 @@
 
 package com.example.android.persistence.db.dao;
 
-
-import android.arch.lifecycle.LiveData;
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.OnConflictStrategy;
-import android.arch.persistence.room.Query;
-
+import androidx.lifecycle.LiveData;
+import androidx.room.Dao;
+import androidx.room.Insert;
+import androidx.room.OnConflictStrategy;
+import androidx.room.Query;
 import com.example.android.persistence.db.entity.CommentEntity;
 
 import java.util.List;
@@ -36,5 +34,5 @@ public interface CommentDao {
     List<CommentEntity> loadCommentsSync(int productId);
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    void insertAll(List<CommentEntity> products);
+    void insertAll(List<CommentEntity> comments);
 }
