@@ -16,12 +16,11 @@
 
 package com.example.android.persistence.db.dao;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.OnConflictStrategy;
-import android.arch.persistence.room.Query;
-
+import androidx.lifecycle.LiveData;
+import androidx.room.Dao;
+import androidx.room.Insert;
+import androidx.room.OnConflictStrategy;
+import androidx.room.Query;
 import com.example.android.persistence.db.entity.ProductEntity;
 
 import java.util.List;
@@ -39,4 +38,8 @@ public interface ProductDao {
 
     @Query("select * from products where id = :productId")
     ProductEntity loadProductSync(int productId);
+
+    @Query("SELECT products.* FROM products JOIN productsFts ON (products.id = productsFts.rowid) "
+        + "WHERE productsFts MATCH :query")
+    LiveData<List<ProductEntity>> searchAllProducts(String query);
 }
