@@ -17,48 +17,33 @@
 package com.example.android.persistence.viewmodel;
 
 import android.app.Application;
-import android.arch.core.util.Function;
-import android.arch.lifecycle.AndroidViewModel;
-import android.arch.lifecycle.LiveData;
-import android.arch.lifecycle.MutableLiveData;
-import android.arch.lifecycle.Transformations;
-
-import com.example.android.persistence.db.DatabaseCreator;
+import androidx.lifecycle.AndroidViewModel;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MediatorLiveData;
+import com.example.android.persistence.BasicApp;
+import com.example.android.persistence.DataRepository;
 import com.example.android.persistence.db.entity.ProductEntity;
-
 import java.util.List;
 
 public class ProductListViewModel extends AndroidViewModel {
 
-    private static final MutableLiveData ABSENT = new MutableLiveData();
-    {
-        //noinspection unchecked
-        ABSENT.setValue(null);
-    }
+    private final DataRepository mRepository;
 
-    private final LiveData<List<ProductEntity>> mObservableProducts;
+    // MediatorLiveData can observe other LiveData objects and react on their emissions.
+    private final MediatorLiveData<List<ProductEntity>> mObservableProducts;
 
     public ProductListViewModel(Application application) {
         super(application);
 
-        final DatabaseCreator databaseCreator = DatabaseCreator.getInstance(this.getApplication());
+        mObservableProducts = new MediatorLiveData<>();
+        // set by default null, until we get data from the database.
+        mObservableProducts.setValue(null);
 
-        LiveData<Boolean> databaseCreated = databaseCreator.isDatabaseCreated();
-        mObservableProducts = Transformations.switchMap(databaseCreated,
-                new Function<Boolean, LiveData<List<ProductEntity>>>() {
-            @Override
-            public LiveData<List<ProductEntity>> apply(Boolean isDbCreated) {
-                if (!Boolean.TRUE.equals(isDbCreated)) { // Not needed here, but watch out for null
-                    //noinspection unchecked
-                    return ABSENT;
-                } else {
-                    //noinspection ConstantConditions
-                    return databaseCreator.getDatabase().productDao().loadAllProducts();
-                }
-            }
-        });
+        mRepository = ((BasicApp) application).getRepository();
+        LiveData<List<ProductEntity>> products = mRepository.getProducts();
 
-        databaseCreator.createDb(this.getApplication());
+        // observe the changes of the products from the database and forward them
+        mObservableProducts.addSource(products, mObservableProducts::setValue);
     }
 
     /**
@@ -66,5 +51,9 @@ public class ProductListViewModel extends AndroidViewModel {
      */
     public LiveData<List<ProductEntity>> getProducts() {
         return mObservableProducts;
+    }
+
+    public LiveData<List<ProductEntity>> searchProducts(String query) {
+        return mRepository.searchProducts(query);
     }
 }
