@@ -16,9 +16,9 @@
 
 package com.example.android.contentprovidersample.data;
 
-import android.arch.persistence.room.ColumnInfo;
-import android.arch.persistence.room.Entity;
-import android.arch.persistence.room.PrimaryKey;
+import androidx.room.ColumnInfo;
+import androidx.room.Entity;
+import androidx.room.PrimaryKey;
 import android.content.ContentValues;
 import android.provider.BaseColumns;
 
