@@ -16,8 +16,7 @@
 
 package com.example.android.persistence.migrations;
 
-import android.arch.persistence.room.TypeConverter;
-
+import androidx.room.TypeConverter;
 import java.util.Date;
 
 public class DateConverter {
