@@ -16,13 +16,12 @@
 
 package com.example.android.persistence.migrations;
 
-import android.arch.persistence.room.ColumnInfo;
-import android.arch.persistence.room.Entity;
-import android.arch.persistence.room.Ignore;
-import android.arch.persistence.room.PrimaryKey;
-
+import androidx.room.ColumnInfo;
+import androidx.room.Entity;
+import androidx.room.Ignore;
+import androidx.room.PrimaryKey;
+import androidx.annotation.NonNull;
 import java.util.Date;
-import java.util.Random;
 import java.util.UUID;
 
 /**
@@ -33,6 +32,7 @@ public class User {
 
     @PrimaryKey
     @ColumnInfo(name = "userid")
+    @NonNull
     private String mId;
 
     @ColumnInfo(name = "username")
